+++ MatriculaAcademica/Controllers/CursosController.cs
@@ -17,28 +17,52 @@ namespace MatriculaAcademica.Controllers
         // GET: Cursos
         public ActionResult Index()
         {
-            return View(db.Curso.ToList());
+            if (Session["tipo"] != null)
+            {
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    return View(db.Curso.ToList());
+                }
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Cursos/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (Session["tipo"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (id == null)
+                    {
+                        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                    }
+                    Curso curso = db.Curso.Find(id);
+                    if (curso == null)
+                    {
+                        return HttpNotFound();
+                    }
+                    return View(curso);
+                }
             }
-            Curso curso = db.Curso.Find(id);
-            if (curso == null)
-            {
-                return HttpNotFound();
-            }
-            return View(curso);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Cursos/Create
         public ActionResult Create()
         {
-            return View();
+            if (Session["tipo"] != null)
+            {
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    return View();
+                }
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Cursos/Create
@@ -48,29 +72,45 @@ namespace MatriculaAcademica.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "id_curso,nome_curso,duracao,turno")] Curso curso)
         {
-            if (ModelState.IsValid)
+            if (Session["tipo"] != null)
             {
-                db.Curso.Add(curso);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (ModelState.IsValid)
+                    {
+                        db.Curso.Add(curso);
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
 
-            return View(curso);
+                    return View(curso);
+                }
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Cursos/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Session["tipo"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (id == null)
+                    {
+                        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                    }
+                    Curso curso = db.Curso.Find(id);
+                    if (curso == null)
+                    {
+                        return HttpNotFound();
+                    }
+                    return View(curso);
+                }
             }
-            Curso curso = db.Curso.Find(id);
-            if (curso == null)
-            {
-                return HttpNotFound();
-            }
-            return View(curso);
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Cursos/Edit/5
@@ -80,28 +120,44 @@ namespace MatriculaAcademica.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "id_curso,nome_curso,duracao,turno")] Curso curso)
         {
-            if (ModelState.IsValid)
+            if (Session["tipo"] != null)
             {
-                db.Entry(curso).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (ModelState.IsValid)
+                    {
+                        db.Entry(curso).State = EntityState.Modified;
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
+                    return View(curso);
+                }
             }
-            return View(curso);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Cursos/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Session["tipo"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (id == null)
+                    {
+                        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                    }
+                    Curso curso = db.Curso.Find(id);
+                    if (curso == null)
+                    {
+                        return HttpNotFound();
+                    }
+                    return View(curso);
+                }
             }
-            Curso curso = db.Curso.Find(id);
-            if (curso == null)
-            {
-                return HttpNotFound();
-            }
-            return View(curso);
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Cursos/Delete/5
@@ -109,10 +165,18 @@ namespace MatriculaAcademica.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Curso curso = db.Curso.Find(id);
-            db.Curso.Remove(curso);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Session["tipo"] != null)
+            {
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    Curso curso = db.Curso.Find(id);
+                    db.Curso.Remove(curso);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         protected override void Dispose(bool disposing)
