+++ MatriculaAcademica/Controllers/CursosDisciplinasController.cs
@@ -17,31 +17,55 @@ namespace MatriculaAcademica.Controllers
         // GET: CursosDisciplinas
         public ActionResult Index()
         {
-            var cursoDisciplina = db.CursoDisciplina.Include(c => c.Curso).Include(c => c.Disciplina);
-            return View(cursoDisciplina.ToList());
+            if (Session["tipo"] != null)
+            {
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    var cursoDisciplina = db.CursoDisciplina.Include(c => c.Curso).Include(c => c.Disciplina);
+                    return View(cursoDisciplina.ToList());
+                }
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: CursosDisciplinas/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (Session["tipo"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (id == null)
+                    {
+                        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                    }
+                    CursoDisciplina cursoDisciplina = db.CursoDisciplina.Find(id);
+                    if (cursoDisciplina == null)
+                    {
+                        return HttpNotFound();
+                    }
+                    return View(cursoDisciplina);
+                }
             }
-            CursoDisciplina cursoDisciplina = db.CursoDisciplina.Find(id);
-            if (cursoDisciplina == null)
-            {
-                return HttpNotFound();
-            }
-            return View(cursoDisciplina);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: CursosDisciplinas/Create
         public ActionResult Create()
         {
-            ViewBag.id_curso = new SelectList(db.Curso, "id_curso", "nome_curso");
-            ViewBag.id_disciplina = new SelectList(db.Disciplina, "id_disciplina", "nome_disciplina");
-            return View();
+            if (Session["tipo"] != null)
+            {
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    ViewBag.id_curso = new SelectList(db.Curso, "id_curso", "nome_curso");
+                    ViewBag.id_disciplina = new SelectList(db.Disciplina, "id_disciplina", "nome_disciplina");
+                    return View();
+                }
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: CursosDisciplinas/Create
@@ -51,33 +75,49 @@ namespace MatriculaAcademica.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "id_curso_disciplina,id_curso,id_disciplina")] CursoDisciplina cursoDisciplina)
         {
-            if (ModelState.IsValid)
+            if (Session["tipo"] != null)
             {
-                db.CursoDisciplina.Add(cursoDisciplina);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (ModelState.IsValid)
+                    {
+                        db.CursoDisciplina.Add(cursoDisciplina);
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
 
-            ViewBag.id_curso = new SelectList(db.Curso, "id_curso", "nome_curso", cursoDisciplina.id_curso);
-            ViewBag.id_disciplina = new SelectList(db.Disciplina, "id_disciplina", "nome_disciplina", cursoDisciplina.id_disciplina);
-            return View(cursoDisciplina);
+                    ViewBag.id_curso = new SelectList(db.Curso, "id_curso", "nome_curso", cursoDisciplina.id_curso);
+                    ViewBag.id_disciplina = new SelectList(db.Disciplina, "id_disciplina", "nome_disciplina", cursoDisciplina.id_disciplina);
+                    return View(cursoDisciplina);
+                }
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: CursosDisciplinas/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Session["tipo"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (id == null)
+                    {
+                        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                    }
+                    CursoDisciplina cursoDisciplina = db.CursoDisciplina.Find(id);
+                    if (cursoDisciplina == null)
+                    {
+                        return HttpNotFound();
+                    }
+                    ViewBag.id_curso = new SelectList(db.Curso, "id_curso", "nome_curso", cursoDisciplina.id_curso);
+                    ViewBag.id_disciplina = new SelectList(db.Disciplina, "id_disciplina", "nome_disciplina", cursoDisciplina.id_disciplina);
+                    return View(cursoDisciplina);
+                }
             }
-            CursoDisciplina cursoDisciplina = db.CursoDisciplina.Find(id);
-            if (cursoDisciplina == null)
-            {
-                return HttpNotFound();
-            }
-            ViewBag.id_curso = new SelectList(db.Curso, "id_curso", "nome_curso", cursoDisciplina.id_curso);
-            ViewBag.id_disciplina = new SelectList(db.Disciplina, "id_disciplina", "nome_disciplina", cursoDisciplina.id_disciplina);
-            return View(cursoDisciplina);
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: CursosDisciplinas/Edit/5
@@ -87,30 +127,46 @@ namespace MatriculaAcademica.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "id_curso_disciplina,id_curso,id_disciplina")] CursoDisciplina cursoDisciplina)
         {
-            if (ModelState.IsValid)
+            if (Session["tipo"] != null)
             {
-                db.Entry(cursoDisciplina).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (ModelState.IsValid)
+                    {
+                        db.Entry(cursoDisciplina).State = EntityState.Modified;
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
+                    ViewBag.id_curso = new SelectList(db.Curso, "id_curso", "nome_curso", cursoDisciplina.id_curso);
+                    ViewBag.id_disciplina = new SelectList(db.Disciplina, "id_disciplina", "nome_disciplina", cursoDisciplina.id_disciplina);
+                    return View(cursoDisciplina);
+                }
             }
-            ViewBag.id_curso = new SelectList(db.Curso, "id_curso", "nome_curso", cursoDisciplina.id_curso);
-            ViewBag.id_disciplina = new SelectList(db.Disciplina, "id_disciplina", "nome_disciplina", cursoDisciplina.id_disciplina);
-            return View(cursoDisciplina);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: CursosDisciplinas/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Session["tipo"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (id == null)
+                    {
+                        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                    }
+                    CursoDisciplina cursoDisciplina = db.CursoDisciplina.Find(id);
+                    if (cursoDisciplina == null)
+                    {
+                        return HttpNotFound();
+                    }
+                    return View(cursoDisciplina);
+                }
             }
-            CursoDisciplina cursoDisciplina = db.CursoDisciplina.Find(id);
-            if (cursoDisciplina == null)
-            {
-                return HttpNotFound();
-            }
-            return View(cursoDisciplina);
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: CursosDisciplinas/Delete/5
@@ -118,10 +174,18 @@ namespace MatriculaAcademica.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            CursoDisciplina cursoDisciplina = db.CursoDisciplina.Find(id);
-            db.CursoDisciplina.Remove(cursoDisciplina);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Session["tipo"] != null)
+            {
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    CursoDisciplina cursoDisciplina = db.CursoDisciplina.Find(id);
+                    db.CursoDisciplina.Remove(cursoDisciplina);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         protected override void Dispose(bool disposing)
