+++ MatriculaAcademica/Controllers/DisciplinasController.cs
@@ -17,28 +17,52 @@ namespace MatriculaAcademica.Controllers
         // GET: Disciplinas
         public ActionResult Index()
         {
-            return View(db.Disciplina.ToList());
+            if (Session["tipo"] != null)
+            {
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    return View(db.Disciplina.ToList());
+                }
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Disciplinas/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (Session["tipo"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (id == null)
+                    {
+                        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                    }
+                    Disciplina disciplina = db.Disciplina.Find(id);
+                    if (disciplina == null)
+                    {
+                        return HttpNotFound();
+                    }
+                    return View(disciplina);
+                }
             }
-            Disciplina disciplina = db.Disciplina.Find(id);
-            if (disciplina == null)
-            {
-                return HttpNotFound();
-            }
-            return View(disciplina);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Disciplinas/Create
         public ActionResult Create()
         {
-            return View();
+            if (Session["tipo"] != null)
+            {
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    return View();
+                }
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Disciplinas/Create
@@ -48,29 +72,45 @@ namespace MatriculaAcademica.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "id_disciplina,nome_disciplina")] Disciplina disciplina)
         {
-            if (ModelState.IsValid)
+            if (Session["tipo"] != null)
             {
-                db.Disciplina.Add(disciplina);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (ModelState.IsValid)
+                    {
+                        db.Disciplina.Add(disciplina);
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
 
-            return View(disciplina);
+                    return View(disciplina);
+                }
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Disciplinas/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Session["tipo"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (id == null)
+                    {
+                        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                    }
+                    Disciplina disciplina = db.Disciplina.Find(id);
+                    if (disciplina == null)
+                    {
+                        return HttpNotFound();
+                    }
+                    return View(disciplina);
+                }
             }
-            Disciplina disciplina = db.Disciplina.Find(id);
-            if (disciplina == null)
-            {
-                return HttpNotFound();
-            }
-            return View(disciplina);
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Disciplinas/Edit/5
@@ -80,28 +120,44 @@ namespace MatriculaAcademica.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "id_disciplina,nome_disciplina")] Disciplina disciplina)
         {
-            if (ModelState.IsValid)
+            if (Session["tipo"] != null)
             {
-                db.Entry(disciplina).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (ModelState.IsValid)
+                    {
+                        db.Entry(disciplina).State = EntityState.Modified;
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
+                    return View(disciplina);
+                }
             }
-            return View(disciplina);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Disciplinas/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Session["tipo"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    if (id == null)
+                    {
+                        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                    }
+                    Disciplina disciplina = db.Disciplina.Find(id);
+                    if (disciplina == null)
+                    {
+                        return HttpNotFound();
+                    }
+                    return View(disciplina);
+                }
             }
-            Disciplina disciplina = db.Disciplina.Find(id);
-            if (disciplina == null)
-            {
-                return HttpNotFound();
-            }
-            return View(disciplina);
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Disciplinas/Delete/5
@@ -109,10 +165,18 @@ namespace MatriculaAcademica.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Disciplina disciplina = db.Disciplina.Find(id);
-            db.Disciplina.Remove(disciplina);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Session["tipo"] != null)
+            {
+                string permissao = (Session["tipo"] as string).Trim();
+                if (string.Equals(permissao, "admin"))
+                {
+                    Disciplina disciplina = db.Disciplina.Find(id);
+                    db.Disciplina.Remove(disciplina);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         protected override void Dispose(bool disposing)
